Move ga-metadata
