+++ frontend/src/metabase-lib/lib/metadata/utils/ga-metadata.js
